+++ crates/rpc/rpc/src/eth/api/pending_block.rs
@@ -50,7 +50,7 @@ impl PendingBlockEnv {
         let block_number = block_env.number.to::<u64>();
 
         let mut executed_txs = Vec::new();
-        let mut best_txs = pool.best_transactions_with_base_fee(base_fee as u128);
+        let mut best_txs = pool.best_transactions_with_base_fee(base_fee);
 
         while let Some(pool_tx) = best_txs.next() {
             // ensure we still have capacity for this transaction
