+++ crates/transaction-pool/src/pool/best.rs
@@ -18,7 +18,7 @@ use tracing::debug;
 /// This iterator guarantees that all transaction it returns satisfy the base fee.
 pub(crate) struct BestTransactionsWithBasefee<T: TransactionOrdering> {
     pub(crate) best: BestTransactions<T>,
-    pub(crate) base_fee: u128,
+    pub(crate) base_fee: u64,
 }
 
 impl<T: TransactionOrdering> crate::traits::BestTransactions for BestTransactionsWithBasefee<T> {
@@ -34,7 +34,7 @@ impl<T: TransactionOrdering> Iterator for BestTransactionsWithBasefee<T> {
         // find the next transaction that satisfies the base fee
         loop {
             let best = self.best.next()?;
-            if best.transaction.max_fee_per_gas() < self.base_fee {
+            if best.transaction.max_fee_per_gas() < self.base_fee as u128 {
                 // tx violates base fee, mark it as invalid and continue
                 crate::traits::BestTransactions::mark_invalid(self, &best);
             } else {
